wtp/webrtc/RTCPeerConnection-track-stats.https.html

<!doctype html>
<meta charset="utf-8">
<title>RTCPeerConnection.prototype.getStats</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script src="dictionary-helper.js"></script>
<script src="RTCStats-helper.js"></script>
<script>
  'use strict';
  
  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   doSignalingHandshake
  //   getUserMediaTracksAndStreams
  
  // The following helper functions are called from RTCStats-helper.js
  // (depends on dictionary-helper.js):
  //   validateRtcStats
  
  var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
  
  async_test(function (t) {
    var pc = new RTCPeerConnection();
    var track = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref) {
      var _ref2 = _slicedToArray(_ref, 2),
          tracks = _ref2[0],
          streams = _ref2[1];
  
      track = tracks[0];
      pc.addTrack(track);
      return pc.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', track.id);
      assert_true(trackStats != null, 'Has stats for track');
      // TODO(hbos): Here and elsewhere, validateRtcStats() only tests id,
      // timestamp and type is correct type. Should validate based on stats type
      // but it expects both audio and video members.
      // https://github.com/w3c/web-platform-tests/issues/9010
      validateRtcStats(report, trackStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'addTrack() without setLocalDescription() yields track stats');
  
  async_test(function (t) {
    var pc = new RTCPeerConnection();
    var stream = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref3) {
      var _ref4 = _slicedToArray(_ref3, 2),
          tracks = _ref4[0],
          streams = _ref4[1];
  
      var track = tracks[0];
      stream = streams[0];
      pc.addTrack(track, stream);
      return pc.getStats();
    })).then(t.step_func(function (report) {
      var streamStats = findStatsByTypeAndId(report, 'stream', stream.id);
      assert_true(streamStats != null, 'Has stats for stream');
      validateRtcStats(report, streamStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'addTrack() without setLocalDescription() yields media stream stats');
  
  async_test(function (t) {
    var pc = new RTCPeerConnection();
    var track = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref5) {
      var _ref6 = _slicedToArray(_ref5, 2),
          tracks = _ref6[0],
          streams = _ref6[1];
  
      track = tracks[0];
      pc.addTrack(track);
      return pc.createOffer();
    })).then(t.step_func(function (offer) {
      return pc.setLocalDescription(offer);
    })).then(t.step_func(function () {
      return pc.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', track.id);
      assert_true(trackStats != null, 'Has stats for track');
      validateRtcStats(report, trackStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'addTrack() with setLocalDescription() yields track stats');
  
  async_test(function (t) {
    var pc = new RTCPeerConnection();
    var stream = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref7) {
      var _ref8 = _slicedToArray(_ref7, 2),
          tracks = _ref8[0],
          streams = _ref8[1];
  
      var track = tracks[0];
      stream = streams[0];
      pc.addTrack(track, stream);
      return pc.createOffer();
    })).then(t.step_func(function (offer) {
      return pc.setLocalDescription(offer);
    })).then(t.step_func(function () {
      return pc.getStats();
    })).then(t.step_func(function (report) {
      var streamStats = findStatsByTypeAndId(report, 'stream', stream.id);
      assert_true(streamStats != null, 'Has stats for stream');
      validateRtcStats(report, streamStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'addTrack() with setLocalDescription() yields media stream stats');
  
  async_test(function (t) {
    var pc = new RTCPeerConnection();
    var track = void 0;
    var stream = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref9) {
      var _ref10 = _slicedToArray(_ref9, 2),
          tracks = _ref10[0],
          streams = _ref10[1];
  
      track = tracks[0];
      stream = streams[0];
      pc.addTrack(track, stream);
      return pc.createOffer();
    })).then(t.step_func(function (offer) {
      return pc.setLocalDescription(offer);
    })).then(t.step_func(function () {
      return pc.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', track.id);
      var streamStats = findStatsByTypeAndId(report, 'stream', stream.id);
      assert_true(trackStats != null && streamStats != null, 'Has stats for track and stream');
      assert_array_equals(streamStats.trackIds, [trackStats.id], 'streamStats.trackIds == [ trackStats.id ]');
      validateRtcStats(report, trackStats);
      validateRtcStats(report, streamStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'addTrack(): Media stream stats references track stats');
  
  // TODO(hbos): addStream() is legacy API not in the spec. Based on discussion
  // whether to standardize in legacy section, consider removing this test or
  // keeping it until addTrack() has wide support.
  // https://github.com/w3c/webrtc-pc/issues/1705
  // https://github.com/w3c/webrtc-pc/issues/1125
  async_test(function (t) {
    var pc = new RTCPeerConnection();
    var track = void 0;
    var stream = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref11) {
      var _ref12 = _slicedToArray(_ref11, 2),
          tracks = _ref12[0],
          streams = _ref12[1];
  
      track = tracks[0];
      stream = streams[0];
      stream.addTrack(track);
      pc.addStream(stream);
      return pc.createOffer();
    })).then(t.step_func(function (offer) {
      return pc.setLocalDescription(offer);
    })).then(t.step_func(function () {
      return pc.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', track.id);
      var streamStats = findStatsByTypeAndId(report, 'stream', stream.id);
      assert_true(trackStats != null && streamStats != null, 'Has stats for track and stream');
      assert_array_equals(streamStats.trackIds, [trackStats.id], 'streamStats.trackIds == [ trackStats.id ]');
      validateRtcStats(report, trackStats);
      validateRtcStats(report, streamStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'Legacy addStream(): Media stream stats references track stats');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    var callee = new RTCPeerConnection();
    var sendingTrack = void 0;
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref13) {
      var _ref14 = _slicedToArray(_ref13, 2),
          tracks = _ref14[0],
          streams = _ref14[1];
  
      sendingTrack = tracks[0];
      caller.addTrack(sendingTrack);
      return doSignalingHandshake(caller, callee);
    })).then(t.step_func(function () {
      return caller.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', sendingTrack.id);
      assert_true(trackStats != null, 'Has stats for sending track');
      var outboundStats = findStatsByTypeAndMember(report, 'outbound-rtp', 'trackId', trackStats.id);
      assert_true(outboundStats != null, 'Has stats for outbound RTP stream');
      validateRtcStats(report, trackStats);
      validateRtcStats(report, outboundStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'O/A exchange yields outbound RTP stream stats for sending track');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    var callee = new RTCPeerConnection();
    var receivingTrack = void 0;
    callee.ontrack = function (trackEvent) {
      assert_true(receivingTrack == undefined, 'ontrack has not fired before');
      receivingTrack = trackEvent.track;
    };
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref15) {
      var _ref16 = _slicedToArray(_ref15, 2),
          tracks = _ref16[0],
          streams = _ref16[1];
  
      caller.addTrack(tracks[0]);
      return doSignalingHandshake(caller, callee);
    })).then(t.step_func(function () {
      return callee.getStats();
    })).then(t.step_func(function (report) {
      assert_true(receivingTrack != null, 'Has a receiving track');
      var trackStats = findStatsByTypeAndId(report, 'track', receivingTrack.id);
      assert_true(trackStats != null, 'Has stats for receiving track');
      var inboundStats = findStatsByTypeAndMember(report, 'inbound-rtp', 'trackId', trackStats.id);
      assert_true(inboundStats != null, 'Has stats for outbound RTP stream');
      validateRtcStats(report, trackStats);
      validateRtcStats(report, inboundStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'O/A exchange yields inbound RTP stream stats for receiving track');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    var callee = new RTCPeerConnection();
    var sendingTrack1 = void 0;
    var sendingTrack2 = void 0;
    var sender = void 0;
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref17) {
      var _ref18 = _slicedToArray(_ref17, 2),
          tracks = _ref18[0],
          streams = _ref18[1];
  
      sendingTrack1 = tracks[0];
      sendingTrack2 = tracks[1];
      sender = caller.addTrack(sendingTrack1);
      return sender.replaceTrack(sendingTrack2);
    })).then(t.step_func(function () {
      return caller.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', sendingTrack2.id);
      assert_true(trackStats != null, 'Has stats for replaced track');
      validateRtcStats(report, trackStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() before offer: new track attachment stats present');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    var callee = new RTCPeerConnection();
    var sendingTrack1 = void 0;
    var sendingTrack2 = void 0;
    var sender = void 0;
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref19) {
      var _ref20 = _slicedToArray(_ref19, 2),
          tracks = _ref20[0],
          streams = _ref20[1];
  
      sendingTrack1 = tracks[0];
      sendingTrack2 = tracks[1];
      sender = caller.addTrack(sendingTrack1);
      return performOffer(caller, callee);
    })).then(t.step_func(function () {
      return sender.replaceTrack(sendingTrack2);
    })).then(t.step_func(function () {
      return caller.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', sendingTrack2.id);
      assert_true(trackStats != null, 'Has stats for replaced track');
      var outboundStats = findStatsByTypeAndMember(report, 'outbound-rtp', 'trackId', trackStats.id);
      assert_true(outboundStats != null, 'Has stats for outbound RTP stream');
      validateRtcStats(report, trackStats);
      validateRtcStats(report, outboundStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() after offer, before answer: new track attachment stats ' + 'present');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    var callee = new RTCPeerConnection();
    var sendingTrack1 = void 0;
    var sendingTrack2 = void 0;
    var sender = void 0;
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref21) {
      var _ref22 = _slicedToArray(_ref21, 2),
          tracks = _ref22[0],
          streams = _ref22[1];
  
      sendingTrack1 = tracks[0];
      sendingTrack2 = tracks[1];
      sender = caller.addTrack(sendingTrack1);
      return doSignalingHandshake(caller, callee);
    })).then(t.step_func(function () {
      return sender.replaceTrack(sendingTrack2);
    })).then(t.step_func(function () {
      return caller.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', sendingTrack2.id);
      assert_true(trackStats != null, 'Has stats for replaced track');
      var outboundStats = findStatsByTypeAndMember(report, 'outbound-rtp', 'trackId', trackStats.id);
      assert_true(outboundStats != null, 'Has stats for outbound RTP stream');
      validateRtcStats(report, trackStats);
      validateRtcStats(report, outboundStats);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() after answer: new track attachment stats present');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    var callee = new RTCPeerConnection();
    var sendingTrack1 = void 0;
    var sendingTrack2 = void 0;
    var sender = void 0;
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref23) {
      var _ref24 = _slicedToArray(_ref23, 2),
          tracks = _ref24[0],
          streams = _ref24[1];
  
      sendingTrack1 = tracks[0];
      sendingTrack2 = tracks[1];
      sender = caller.addTrack(sendingTrack1);
      return doSignalingHandshake(caller, callee);
    })).then(t.step_func(function () {
      return sender.replaceTrack(sendingTrack2);
    })).then(t.step_func(function () {
      return caller.getStats();
    })).then(t.step_func(function (report) {
      var trackStats = findStatsByTypeAndId(report, 'track', sendingTrack1.id);
      assert_true(trackStats != null, 'Has stats for original track');
      assert_true(trackStats.objectDeleted);
      var outboundStats = findStatsByTypeAndMember(report, 'outbound-rtp', 'trackId', trackStats.id);
      assert_true(outboundStats == null, 'The outbound RTP stream should no longer reference the ' + 'original attachment');
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack(): original track attachment stats present after replacing');
  
  // Helpers.
  
  function findStatsByTypeAndId(report, type, identifier) {
    return findStats(report, function (stats) {
      return stats.type == type && stats[type + 'Identifier'] == identifier;
    });
  }
  
  function findStatsByTypeAndMember(report, type, member, value) {
    return findStats(report, function (stats) {
      return stats.type == type && stats[member] == value;
    });
  }
  
  function findStats(report, findFunc) {
    for (var it = report.values(), n = it.next(); !n.done; n = it.next()) {
      if (findFunc(n.value)) return n.value;
    }
    return null;
  }

</script>
